+++ cbindgen.toml
@@ -1,3 +1,3 @@
 language = "C++"
 namespace = "crustyArchitecture"
-include_guard = "crustyArchitecture.HPP"
+include_guard = "CRUSTY_ARCHITECTURE_HPP"
